requirements.txt
pandas
psycopg2
pandas_cohort
python-dotenv
sqlalchemy
memorize.py